requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyyaml>=6.0
typer>=0.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
openai>=1.0.0
tqdm>=4.64.0

# Development tools
ruff>=0.1.0
pre-commit>=3.5.0
